Group_Constitution_v1
This is the template for our group's constitution, provided by ULife. It will be required when applying for ULife recognition, which will make all subsequent group affiliation requests way easier. We should aim to have this document up and ready to go by the second half of July, so fill in/ or comment on anything that you think should be added or changed!