add ways to get involved + toc, coc, constitution
As per https://github.com/UofTCoders/studyGroup/issues/258
Please feel free to edit and add things I forgot. Just wanted to get a lot of the basics down, and feel free to refine as needed.
Also not really sure if the links in the Table of Contents will work, but headers are automatically anchors right?